refactor(DSWx-NI): moved gcov file to gcov package
